Move some utiliy functions to the utils file
Easier to put them there
